[Core] remove some legacy header files
