+++ src/inference/dev_api/cpp_interfaces/interface/ie_ivariable_state_internal.hpp
@@ -8,7 +8,7 @@
 #include <string>
 
 #include "ie_blob.h"
-#include "so_ptr.hpp"
+#include "openvino/runtime/so_ptr.hpp"
 
 namespace InferenceEngine {
