--- src/inference/dev_api/so_ptr.hpp
@@ -1,11 +1,0 @@
-// Copyright (C) 2018-2023 Intel Corporation
-// SPDX-License-Identifier: Apache-2.0
-//
-
-/**
- * @brief This is a wrapper class for handling plugin instantiation and releasing resources
- * @file so_ptr.hpp
- */
-#pragma once
-
-#include "openvino/runtime/so_ptr.hpp"
