+++ src/inference/src/cpp_interfaces/interface/ie_iplugin_internal.cpp
@@ -24,7 +24,6 @@
 #include "cpp/ie_cnn_network.h"
 #include "dev/converter_utils.hpp"
 #include "exec_graph_info.hpp"
-#include "ie_algorithm.hpp"
 #include "ie_api.h"
 #include "ie_icore.hpp"
 #include "ie_iextension.h"
@@ -274,7 +273,7 @@ std::unordered_set<std::string> GetRemovedNodes(const std::shared_ptr<const ov::
     }
 
     for (auto&& originalNode : originalFunction->get_ops()) {
-        if (!InferenceEngine::details::contains(transformedNodeNames, originalNode->get_friendly_name()))
+        if (transformedNodeNames.find(originalNode->get_friendly_name()) == transformedNodeNames.end())
             result.emplace(originalNode->get_friendly_name());
     }
