--- src/inference/tests/functional/debug_tests.cpp
@@ -1,21 +1,0 @@
-// Copyright (C) 2018-2023 Intel Corporation
-// SPDX-License-Identifier: Apache-2.0
-//
-
-#include <gtest/gtest.h>
-
-#include <string>
-
-#include "debug.h"
-
-using DebugTests = ::testing::Test;
-
-TEST_F(DebugTests, tolowerWorksWithEmptyString) {
-    std::string str = "";
-    ASSERT_STREQ("", InferenceEngine::details::tolower(str).c_str());
-}
-
-TEST_F(DebugTests, shouldConvertToLowerCase) {
-    std::string str = "Hello, World!1";
-    ASSERT_STREQ("hello, world!1", InferenceEngine::details::tolower(str).c_str());
-}
