+++ src/plugins/intel_cpu/src/infer_request.cpp
@@ -6,7 +6,6 @@
 
 #include "async_infer_request.h"
 #include "compiled_model.h"
-#include "debug.h"
 #include "dnnl_extension_utils.h"
 #include "itt.h"
 #include "memory_desc/dnnl_blocked_memory_desc.h"
