+++ src/tests/functional/plugin/shared/src/behavior/plugin/hetero_synthetic.cpp
@@ -9,7 +9,6 @@
 #include "common_test_utils/file_utils.hpp"
 #include "openvino/util/file_util.hpp"
 #include <random>
-#include "ie_algorithm.hpp"
 #include "common_test_utils/subgraph_builders/split_conv_concat.hpp"
 #include "common_test_utils/subgraph_builders/split_multi_conv_concat.hpp"
 #include "common_test_utils/subgraph_builders/nested_branch_conv_concat.hpp"
